regression-cmr/requirements.txt
torch==1.10
SimpleITK==2.0.2
torchvision
torchio
pandas
numpy
neptune-client
nibabel
napari
matplotlib
seaborn
kornia
unet
tqdm
self_attention_cv
optiocalflow3d
itk-elastix